% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_munis.R
\name{clean_munis}
\alias{clean_munis}
\title{Clean the municipality list}
\usage{
clean_munis(df, muni_status, muni_var, time_var = NULL, vars_to_clean)
}
\arguments{
\item{df}{Provide a data frame you want to clean}

\item{muni_status}{Use the output from `mm::which_muni_status` here or define the status by giving a "yyyy-mm-dd" date}

\item{muni_var}{Define the variable name containing the municipal ID's}

\item{time_var}{Define the variable name of the time dimension (if left empty, a cross-section is assumed)}

\item{vars_to_clean}{Define the variable name or vector of variable names of variables you need to aggregate}
}
\value{
A data frame containing the clean municipality list
}
\description{
Clean the municipality list
}
\examples{
muni_status <- mm::which_muni_status(df = vacancy_stock, target_year = 2018)
mm::clean_munis(df = vacancy_stock, muni_status = muni_status, muni_var = "commune_id", time_var = "year", vars_to_clean = c("vacant", "stock"))
}
