% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_quality.R
\name{clean_quality}
\alias{clean_quality}
\title{Performs a final check of cleaning quality and overwrites the original data frame}
\usage{
clean_quality(df, clean_df, muni_status, muni_var, time_var = NULL)
}
\arguments{
\item{df}{Provide the target data frame to be over-written with the cleaned version}

\item{clean_df}{Provide the clean data frame from `mm::clean_munis`}

\item{muni_status}{Use the output from `mm::which_muni_status` here}

\item{muni_var}{Define the variable name containing the municipal ID's}

\item{time_var}{Define the variable name of the time dimension (if left empty, a cross-section is assumed)}
}
\value{
The overwritten clean data frame
}
\description{
Performs a final check of cleaning quality and overwrites the original data frame
}
\examples{
muni_status <- mm::which_muni_status(df = vacancy_stock, target_year = 2018)
df_clean <- mm::clean_munis(df = vacancy_stock, muni_status = muni_status, muni_var = "commune_id", time_var = "year", vars_to_clean = c("vacant", "stock"))
mm::clean_quality(df = vacancy_stock, clean_df = clean_df, muni_status = muni_status, muni_var = "commune_id", time_var = "year")
}
