% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_to_lv95.R
\name{gps_to_lv95}
\alias{gps_to_lv95}
\title{GPS to LV95}
\usage{
gps_to_lv95(long, lat)
}
\arguments{
\item{long}{Longitude, in degrees.}

\item{lat}{Latitude, in degrees.}
}
\value{
Named list with \code{long} and \code{lat}.
}
\description{
Transforms GPS coordinates to Swiss LV95 coordinates, such that they can be used with the BFS maps.
}
\examples{
gps_to_lv95(long = 7.45, lat = 46.95)
}
\author{
Tino Good, \email{onit.good+suistats@gmail.com}
}
