% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_layer_data.R
\name{make_layer_data}
\alias{make_layer_data}
\title{Create Map-Layers for ggplot2}
\usage{
make_layer_data(agg_level = "GDNR")
}
\arguments{
\item{agg_level}{Default is BFS-ID (GMDNR), Choose one of "GMDNR", "KTNR", "BZNR", "MSNR", "GRNR".}
}
\value{
A fortified dataframe that is ready for use with ggplot2.
}
\description{
Create a Swiss map for ggplot2 with desired borders/levels.
}
\examples{
make_layer_data(agg_level = "GMDNR")
}
