% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapvalues.R
\name{mapvalues}
\alias{mapvalues}
\title{Map values}
\usage{
mapvalues(x, from, to, warn_missing = FALSE, missing_to_na = TRUE)
}
\arguments{
\item{x}{Vector to be translated/recode.}

\item{from}{From-values.}

\item{to}{To-values. Note: same order and length as from!}

\item{warn_missing}{Warning, if missings occur (not every in \code{x} has a match in \code{from}). Default is \code{FALSE}.}

\item{missing_to_na}{If \code{TRUE} (default), values that do not occur in the from-vector will be transformed to \code{NA}. If \code{FALSE}, these values will remain as given in \code{x}, hence untouched.}
}
\value{
Vector of length \code{x} with mapped values.
}
\description{
Maps values of given input to another value, given the vectors \code{from} and \code{to}. This function is a modified version of \code{plyr::mapvalues}, but treats missing values more strict (by default).
}
\examples{
mapvalues(x = letters, from = letters, to = rev(letters)) # recode letters from a-z by reverse order, thus output is z-a
}
