% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_level.R
\name{switch_level}
\alias{switch_level}
\title{Switch from GDNR to other BFS-levels}
\usage{
switch_level(x, from = "GDNR", to, ...)
}
\arguments{
\item{x}{Vector with values to be mapped.}

\item{to}{Output category.}

\item{...}{Pass other settings to \code{mapvalues}.}
}
\value{
Vector of same length as input with the respective to-values.
}
\description{
Maps values from municipality ID (\code{GDNR}) one to another, including zip-codes, BFS-IDs, Grossraumnummer,
... check \code{names(bfs_regio_info)} for alle options.
}
\examples{
switch_level(x = 1000, to = "MSNR")
}
\author{
Tino Good, \email{onit.good+suistats@gmail.com}
}
