% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_bfs_id.R
\name{update_bfs_id}
\alias{update_bfs_id}
\title{Merge Municipalities}
\usage{
update_bfs_id(bfs_id, target_year = Inf)
}
\arguments{
\item{bfs_id}{Vector with BFS IDs.}

\item{target_year}{Target year. Default ist most recent available (therefore \code{Inf}).}
}
\value{
Vector with BFS IDs as of \code{target_year}.
}
\description{
Merge municipalities af of a certain target year to enable combination with statistics or geodata as of target year (including maps etc.).
}
\examples{
update_bfs_id(bfs_id = 132, target_year = 2017) # still 132, since merger only happens in 2018:
update_bfs_id(bfs_id = 132, target_year = 2018) # now 295
update_bfs_id(bfs_id = 132) # default: most recent
}
\author{
Tino Good, \email{onit.good+suistats@gmail.com}
}
