% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_muni_status.R
\name{which_muni_status}
\alias{which_muni_status}
\title{Check which municipality list to use for a given target year}
\usage{
which_muni_status(df, target_year, muni_var, time_var = NULL)
}
\arguments{
\item{df}{Define the data frame you wish to clean that contains at least a municipality id column (optional is an annual time dimension column)}

\item{target_year}{Define the target year which determines the final list of municipalities}

\item{muni_var}{Provide the variable name that identifies the municipal ID's}

\item{time_var}{Provide the variable name for the time dimesion (if left empty, a cross-section is assumed)}
}
\value{
A string indicating which municipality list definition to use, for a given target year
}
\description{
Check which municipality list to use for a given target year
}
\examples{
mm::which_muni_status(df = vacancy_stock, target_year = 2018)
}
