% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plz4_to_gdnr.R
\name{plz4_to_gdnr}
\alias{plz4_to_gdnr}
\title{Transform PLZ4 to GDNR}
\usage{
plz4_to_gdnr(x, true_match_only = FALSE)
}
\arguments{
\item{x}{Vector with PLZ4.}
}
\value{
Vector with municipality numbers according to BFS (GDNR).
}
\description{
Transform PLZ4 to GDNR
}
